// src/store/services/types/messages.ts
export enum MessagePartType {
  TEXT = 0,
  MENTION = 4,
  LINK = 5,
  TWITCH_EMOTE = 6,
  TWITCH_CLIP = 7,
  TWITCH_VIDEO = 8,

  BTTV_EMOTE = 101,
  FFZ_EMOTE = 102,
  STV_EMOTE = 103,
  EMOJI = 104,
}

export const IRCV3_KNOWN_COMMANDS = new Map([['PRIVMSG', PrivateMessage]]);

export interface MessagePartText {
  type: MessagePartType.TEXT;
  content: string;
}

interface MessagePartEmoteContent {
  id: string;
  modifiers: MessagePartEmoteModifier[];
}

// twitch message
export interface MessagePartTwitchEmote {
  type: MessagePartType.TWITCH_EMOTE;
  content: MessagePartEmoteContent;
}

// bttv emote
export interface MessagePartBttvEmote {
  type: MessagePartType.BTTV_EMOTE;
  content: MessagePartEmoteContent;
}

// ffz emote
export interface MessagePartFfzEmote {
  type: MessagePartType.FFZ_EMOTE;
  content: MessagePartEmoteContent;
}

// seven tv emote
export type MessagePartStvEmote = {
  type: MessagePartType.STV_EMOTE;
  content: MessagePartEmoteContent;
};

// emoji emote
export type MessagePartEmoji = {
  type: MessagePartType.EMOJI;
  content: MessagePartEmoteContent;
};

// mention
export interface MessagePartMention {
  type: MessagePartType.MENTION;
  content: {
    displayText: string;
    recipient: string;
  };
}

// link
export interface MessagePartLink {
  type: MessagePartType.LINK;
  content: {
    displayText: string;
    url: string;
  };
}

// twitch clip
export type MessagePartTwitchClip = {
  type: MessagePartType.TWITCH_CLIP;
  content: {
    displayText: string;
    slug: string;
    url: string;
  };
};

// twitch video
export interface MessagePartTwitchVideo {
  type: MessagePartType.TWITCH_VIDEO;
  content: {
    displayText: string;
    id: string;
    url: string;
  };
}

export type MessagePart =
  | MessagePartText
  | MessagePartTwitchEmote
  | MessagePartBttvEmote
  | MessagePartFfzEmote
  | MessagePartStvEmote
  | MessagePartEmoji
  | MessagePartMention
  | MessagePartLink
  | MessagePartTwitchClip
  | MessagePartTwitchVideo;
// ...
export enum MessageCardType {
  TWITCH_CLIP = 0,
  TWITCH_VIDEO = 1,
  YOUTUBE_VIDEO = 2,
}

// https://regex101.com/r/jGbDV1/5
export const TWITCH_CLIP_REGEX =
  /^(?:https?:\/\/)?(?:clips\.twitch\.tv\/|(?:www\.|m\.)?twitch\.tv\/(?:[\d\w]+)\/clip\/)([\d\w-]+)(?:\?.+)?$/;

// https://regex101.com/r/xsgeA4/4
export const TWITCH_VIDEO_REGEX =
  /^(?:https?:\/\/)?(?:www\.|m\.)?twitch\.tv\/videos\/(\d+)(?:\?.+)?$/;

// https://regex101.com/r/22dNgk/1
export const YOUTUBE_VIDEO_REGEX =
  /^((?:https?:)?\/\/)?((?:www|m)\.)?(?:youtube\.com|youtu.be)(\/(?!channel)(?:[\w-]+\?v=|embed\/|v\/|shorts\/)?)([\w-]+)(\S+)?$/;

export type MessageCard = {
  type: MessageCardType;
  id: string;
  url: string;
};